requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.5
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2
# Gemini Python SDK only (Gemini-only build)
google-generativeai>=0.7.0
pypdf==4.3.1
python-docx==1.1.2
python-multipart==0.0.9
twilio==9.3.2